✅ Minimum Number of Days to Make m Bouquets | Medium | Java | Runtime: 16 ms (91.58%) | Memory: 82.6 MB | 2026-07-06

+++ leetcode/medium/1482-minimum-number-of-days-to-make-m-bouquets/solution.java
@@ -13,7 +13,7 @@ class Solution {
         //m1
         //int ans =-1 ;
         //m2
-        if (k*m > bloomDay.length){
+        if ((long)k*m > bloomDay.length){
             return -1;
         }
